+++ application/dash_application/views/call_ready_wgs.py
@@ -57,8 +57,8 @@ ids = init_ids([
     'all-count',
 ])
 
-BAMQC_COL = gsiqcetl.column.BamQc4MergedColumn
-CALL_COL = gsiqcetl.column.MutetctCallabilityColumn
+BAMQC_COL = qcetl.column.BamQc4MergedColumn
+CALL_COL = qcetl.column.MutetctCallabilityColumn
 PINERY_COL = pinery.column.SampleProvenanceColumn